refactor: porfolio moved from AboutMe
	modified:   src/components/AboutMe/AboutMe.css
	modified:   src/components/AboutMe/AboutMe.js
