+++ src/components/AboutMe/AboutMe.js
@@ -14,21 +14,6 @@ const AboutMe = () => {
     },
   ];
 
-  const portfolioLinks = [
-    {
-      name: "Cтатичный сайт",
-      url: "https://",
-    },
-    {
-      name: "Адаптивный сайт",
-      url: "https://",
-    },
-    {
-      name: "Одностраничное приложение",
-      url: "https://",
-    },
-  ];
-
   return (
     <section className="student main-sizer">
       <SectionHeading text="Студент" />
@@ -60,23 +45,6 @@ const AboutMe = () => {
             })}
           </ul>
         </div>
-      </div>
-      <div className="portfolio">
-        <h4 className="portfolio__heading">Портфолио</h4>
-        <ul className="portfolio__links">
-          {portfolioLinks.map((link, index) => {
-            return (
-              <li className="portfolio__link" key={index}>
-                <a className="portfolio__url" href={link.url}>
-                  {link.name}
-                </a>
-                <a className="portfolio__arrow" href={link.url}>
-                  {""}
-                </a>
-              </li>
-            );
-          })}
-        </ul>
       </div>
     </section>
   );
