+++ src/components/Main/Main.js
@@ -3,6 +3,7 @@ import Promo from "../Promo/Promo";
 import AboutProject from "../AboutProject/AboutProject";
 import Techs from "../Techs/Techs";
 import AboutMe from "../AboutMe/AboutMe";
+import Portfolio from "../Portfolio/Portfolio";
 
 const Main = () => {
   return (
@@ -11,6 +12,7 @@ const Main = () => {
       <AboutProject />
       <Techs />
       <AboutMe />
+      <Portfolio />
     </main>
   );
 };
